Fix string deallocation issues
DBToaster is not Gollum, and RAM is not a ring,
give back the memory and deallocate the thing.
It's unclear why destruction in class PString was disabled,
since the commit messages were not sufficiently labelled.

Grad students (at some point) learn that constuctors are hard,
so let's be smart, and use std::string from the start.

(on a technical note, the first path to fixing PString
is to re-enable the destructor, but the code has
more issues. functional correctness is more important
in our scenario than few percent of performance gain
for some specific use-cases)

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>
Partly-rhymed-by: Ralf Ramsauer <ralf.ramsauer@othr.de>