Map input data directly into memory
The purpose of reading the input data from a stream into
a buffer is to avoid I/O noise during the measurement proper.
However, this method does not guarantee intervention by the
kernel, and can be done more reliably by mapping the input
data (on Linux) with MAP_POPULATE. If allocations are properly
lockes, this guarantees that the data will remain in RAM
until deallocated.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>